Bug#17590161 READ/WRITE MUTEXES ON BINLOG DELEGATE CLASSES ARE NOT COUNTED ON PERF SCHEMA
Prior to this fix, the read write lock Delegate::lock was not instrumented 
for the performance schema.
This lock is used in particular for the following classes:
- Trans_delegate
- Binlog_storage_delegate
- Binlog_transmit_delegate
- Binlog_relay_IO_delegate

With this fix, the read write lock is instrumented.
For clarity, a different name is used for each sub class,
to have distinct statistics for distinct uses.

The instruments added are visible in table 
performance_schema.setup_instruments as:
- wait/synch/rwlock/sql/Trans_delegate::lock
- wait/synch/rwlock/sql/Binlog_storage_delegate::lock
- wait/synch/rwlock/sql/Binlog_transmit_delegate::lock
- wait/synch/rwlock/sql/Binlog_relay_IO_delegate::lock