Bug#16884217 BTR_VALIDATE_LEVEL() FAILS ON THE CHANGE BUFFER TREE
ibuf_restore_pos(): If the change buffer cursor restoration fails,
do not invoke btr_validate_index() on the change buffer, but instead
crash as fast as possible, so that we would have more useful information
in the core dump.

When WL#5522 (transportable tablespaces) was pushed, it added checks to
btr_validate_level(), checking that each page is marked as allocated.
Because the allocation of change buffer pages is handled differently from
other B-tree pages, these checks would fail on the change buffer.

This was the only call to btr_validate_index() on the change buffer.
Normally, btr_validate_index() is invoked for indexes of user tables,
as part of CHECK TABLE.

Approved by Jimmy Yang over IM.