A page in InnoDB is identified by a tablespace number (called space or space_id in the code) and a page number within that tablespace (called page_no, offset or page in the code). Often we also use zip_size along with space,page_no to designate the compressed page size within that tablespace although it can be retrieved, knowing the space id. Another parameter that is also used often with space,page_no,zip_size is the fold value, derived from space,page_no used in some hashing functions.
In this patch we introduce a new class that represents a page
identifier - it contains space,page_no and also the derived values
zip_size and fold to avoid multiple recalculation or retrieval of the
latter. The main purpose of this patch is to simplify the code - e.g.
instead of passing 3 parameters to functions space,page_no,zip_size
pass just one and to optimize the fold value recalculation.