WL#6797 Method for clearing session state
As of today connection pools partitioned by user use COM_CHANGE_USER during 
recycle of a connection, by pool manager which does a re-authentication which
is not needed. This WL will introduce a new RPC COM_RESET_CONNECTION which 
will only do the cleanup of the session context without re-authentication.