Bug#16924125: LOG_SLOW_STATEMENT COULD BAIL OUT EARLIER IF               SLOW LOGGING IS DISABLED
The problem was that checking if the slow query log was enabled,
was one of the last checks that was done before writing to the
slow query log. Since the slow query log by default is disabled,
it makes sense to have this check earlier.

This patch moves the check from Query_logger::slow_log_write() to
the start of log_slow_applicable().

As this is a minor optimization with no changes of behavior,
no test case is added.

The patch also fixes the --log-throttle-queries-not-using-indexes
startup option so that it requires a value (=# instead of [=#]).