Bug#17303649: INCORRECT COST COMPARISON IN FIND_BEST_REF()
The cost comparison at the end of find_best_ref() is incorrect. 
It uses 
"read cost of ref access for all rows from earlier tables" +
"CPU cost of handling rows returned from a single ref access"

Thus, the read cost and the CPU cost is not comparable. The CPU
cost also needs to be "for all rows from earlier tables".