WL6797: Method for clearing session state.
Probelm: main.mysql_upgrade tests fails on rhel5-x86-64bit-valgrind platform. 
Fix: As part of WL6797 in main.mysql i have introduced a testcase which does not 
do a cleanup of a table wl6797.t1, because of which there is a result content
mismatch in mysqlcheck.result file. The issue is fixed by doing a proper cleanup.
This is a test file issue.