BUG#17590616:CAN'T USE LOGICAL_CLOCK IF GTID IS ENABLED
Problem: When GTID is used with Intra-schema
multi-threaded slave the transactions are only
assigned to the first worker, thereby compromizing
the multi-threaded nature of replication slave.
      
Background: Master stores the commit parent in the
first query of a transaction (i.e. BEGIN query) if
GTID is disabled or in GTID event if GTID is enabled.
But in the slave we called the scheduler method for
both the GTID event and the begin event. But since in
case of GTID the BEGIN event does not have the commit
parent, the scheduler logic blocks the execution of the
event until the previous transaction is applied completely.
This causes the slave coordinator to always assign the event
to the first worker only.
      
Fix: We fix this problem by calling the scheduler method for
the "BEGIN" only when the coordinator has not seen the GTID
event already. This ensures that it is never called for the
first event of the transaction when GTID is enabled.