Bug#17500866 -	SELECT COUNT(DISTINCT ..) RETURNS WRONG RESULT WHEN                 TEMP TABLE SIZE IS LOW
Problem: 
When running a simple COUNT(DISTINCT) query on a table that 
contains a large number of distinct values, the query can 
return a wrong result. In most cases the query will return 0 
or an empty result. This is most likely to happen when the 
maximum size of internal temporary tables are set to a 
low value by setting either tmp_table_size or 
max_heap_table_size to a low value. 

Analysis:
Suppose the table contains records where the field for doing 
COUNT(DISTINCT) on has the following values:


-->1 1 9 5 4 1 0 8 11 7 12 3 6

Assume that tmp_table_size is set to 16 in this example. 

First we find the first 4 unique value and put it is a file in a sorted
order (why 4? because size of field is 4 and tmp_table_size is 16)
that is 1 4 5 9-->f1
then second
0 1 8 11-->f2
third will have
3 6 7 12-->f3

NOTE:f1,f2,f3 are pointing to same file
after fetching whole record we

Doing the COUNT(DISTINCT) is done by Unique::walk() which in this case 
will call merge_walk() to do the main job. In merge_walk() the first 
elements from each file: 

f1--> 1 4
f2--> 0 1
f3--> 3 6

here we compare the first value, which ever is least remove 
it(and all other which have same value from the queue and 
increment the count distinct counter.

like in this we will remove 0 from f2.then we have like
f1--> 1 4
f2--> 1
f3--> 3 6
then we will remove 1 from f1, f2  and fetch remaining value for f2
it will be like this
f1--> 4
f2--> 8 11
f3--> 3 6

Like this we will get count distinct value.

When counting the distinct values in merge_walk() a merge buffer 
is used. This buffer must be able to store one record from each 
of the file pointer (f1, f2, f3 above). The merge_buffer is allocated 
in Unique::walk() and  with a size that is equal to 
MIN(tmp_table_size, max_heap_table_size). When tmp_table_size or 
max_heap_table_size is very small, there might be a large number of file 
pointers to read from. In the case where this bug occurs, the number of 
files is so large that there is not enough space to store one record 
from each file pointer. This situation is detected in merge_walk() and 
handled by just returning an error status. The COUNT(DISTINCT) operation 
is not done and this causes 0 to be returned as the result.


Solution:
Try to allocate the merge buffer with a size where it have space to store 
atleast one record from each file pointers.

Note:
Till now we are not taking care of the cases, where situation is same but 
we can't able to allocate that much memory. In those cases again we will give 
result as zero.