WL#6120- Change master without stopping slave
Currently, a CHANGE MASTER requires a STOP SLAVE
to kill the IO and SQL threads. Also, In order
to switch the I/O thread over to read from another
master it is currently necessary to stop the SQL
thread as well.
      
This worklog implement support for
1) Not having to STOP SLAVE, wherever possible.
2) Re-directing the I/O thread without having to
   stop the SQL thread.
3) Allow change Master without stopping IO thread
   wherever possible

This patch completes the coding except for some loose
ends.