Bug#17560636: FIND_BEST_REF(): COST FOR CLUSTERED PK IS               CALCULATED AS IF NOT COVERING
Optimize_table_order::find_best_ref() iterates through all 
possible ways to access all usable keys for a table. The index 
access strategy with lowest cost is chosen. However, if a 
clustered primary key is applicable, it is considered to be 
covering only if all columns required for the query are in
the primary key definition. The result is that the cost for 
clustered primary keys is calculated as though a random read 
needs to take place for each estimated row. This is incorrect 
when clustered primary keys are read since the data can be 
read from the index without going to "the table".

The fix is to let handler::read_time() calculate the cost 
instead of assuming that there will be a cost of 1 x #row. 
handler::read_time() is capable of handling the distinction
between reading the clustered index and another index.