Fix for Bug#17154722 REMOVE COMPILER WARNINGS PRODUCED BY SUN STUDIO COMPILER
Fixes the following compiler warning caused by a local variable having the
same name as a class variable:
      
   Warning: keyuse hides JOIN::keyuse
      
The solution for this is to rename the JOIN::keyuse variable to keyuse_array.
The patch also changes make_join_statistics() and optimize_keyuse() to use 
JOIN::keyuse_array directly instead of getting it as an argument.