cmake - ensure only one rule produces sql_yacc.cc and sql_yacc.h
If one builds out-of-source from a source distribution that includes
sql_yacc.cc and sql_yacc.h, there was two rules producing those in
the build tree.

One copying the files from source and a second that run bison to generate them.

This patch makes sures that files are copied only if bison was not used.

This should eliminate the risk that some part of make uses the copied files
while bison generates new ones.