Bug#17695635: GUNIT TESTS CANNOT BE COMPILED WITH LIBC++
GMock 1.6 assumes that <tr1/tuple> exists. But this is not true for libc++
where tuple is in std. Added CMake check for this which makes GMock
use it's own tuple implementation if <tr1/tuple> does't exist.

Also changed stdcxx-t.cc to look for unordered_map in <unordered_map>
rather than <tri/unordered_map> when using libc++.