Bug#17713031: Assert 'table == tables[unique_counter] in multi_delete
This is a regression from the refactoring of multi_delete in WL#6987.

The arrays tempfiles and tables in class multi_delete have one entry
for each table to be deleted from, which is not deleted from immediately.
(The deletes are collected inside a Unique object and deletes happen
after the "immediate" deletes).

However, the code that incremented the counter unique_counter which
keeps track of the tempfile object for the current table being deleted
from was placed in a wrong location, so that when a row from an
outer-joined table was NULL-complemented, unique_counter was not
incremented as it should.

The fix is to place the increment operation earlier, just after
detecting that we have a table that is deleted from, and the delete
is not immediate. This placement of the code will also prevent other
conditionals from interfering with this logic (such as e.g. an error
in a triggered operation).