Bug#17720496: Test failure in innodb.innodb_mysql with query cache
This is a regression after WL#6987.

The problem can be reproduced with the following test:

./mysql-test-run --mem main.view \
--mysqld=--query_cache_type=1 --mysqld=--query_cache_size=1M

The problem was a missing rewrite of a query cache update for LOAD
statements, so that the query cache assumed a view was updated and
not the underlying base table.

The fix is to use the same query cache invalidate function as
INSERT does.