Bug#13944462 'NULL IN (XX)' RETURNS WRONG RESULTS
A IN (B1, B2, ...) can be evaluated with two different methods:
1) we store the values of Bs into an array, sort the array, and search
for A using bisection 
2) or we simply evaluate like if it were "A=B1 OR A=B2 OR etc".
The function which runs one of those two methods is
Item_func_in::val_int() (the first method is the "if (array)" section
there).

In our bug report, A and Bs are themselves tuples:
select (NULL, 1, NULL, 2) IN ((0, 0, 0, 0), (0, 0, 0, 0));

The function which chooses which method to run is
Item_func_in::fix_length_and_dec(). The part of it which is of
interest is:
  if (type_cnt == 1)  <<<<<< A and Bs all of same type...
  {
    /*
      When comparing rows create the row comparator object beforehand to ease
      the DATETIME comparison detection procedure.
    */
    if (cmp_type == ROW_RESULT) <<<<<< ... which type is ROW
    {
      cmp_item_row *cmp= 0;
      if (const_itm && !nulls_in_row())
      {
        array= new in_row(arg_count-1, 0);

const_itm is true if all Bs are made of constants. nulls_in_row () is
true if no B contains a NULL. In our bug report, both are true, so we
create the array and use bisection.
This decision is incorrect. Indeed, it misses that A contains a
NULL. Suppose that we are evaluating:
(NULL,1) IN ((1,2),(2,2),(3,C).
The rules of SQL say that the result should be UNKNOWN if C=1, and
FALSE otherwise. To be able to determine this result, MySQL has to
look at the third row. That is not possible with bisection in the
lexicographically sorted array: first we would need to compare
(NULL,1) with the middle element of the array: (2,2). This comparison
must return one of "greater, smaller, equal"; how to do this when one
argument is NULL? Even if we choose to ignore NULL, that is, we
compare only the second elements of the tuples (1 versus 2), that
would direct us into the first half of the array, so we would never
meet (3,C).

It is for similar reasons that when the right argument of IN is a
subquery, we cannot use ref access into the subquery's table if the
left argument is NULL.

See also this existing comment:
  /*
    Row item with NULLs inside can return NULL or FALSE =>
    they can't be processed as static
  */

First fix: if it really matters to distinguish between
UNKNOWN and FALSE (i.e. IN is not top-level), and the left argument
contains NULL, don't do bisection in a sorted array: choose method 2).
For that, I implement Item_func_in::top_level_item() and
Item_func_in::is_top_level_item(). I also rename
Item_func_not_all:top_level() to is_top_level_item(), to make the name
similar to Item_subselect::is_top_level_item().

After this first fix, our testcase starts using method 2).
Unfortunately, the query's result is still incorrect. Indeed, that
method executes:
    if (!(value_added_map & (1U << (uint)cmp_type)))
    {
      in_item->store_value(args[0]);
      if ((null_value= args[0]->null_value)) <<<<<<<<<<<
        return 0;
      value_added_map|= 1U << (uint)cmp_type;
    }
    if (!in_item->cmp(args[i]) && !args[i]->null_value)  <<<<<<<<
      return (longlong) (!negated);
    have_null|= args[i]->null_value;
  }
  null_value= have_null;
args[0] is the left argument of IN (a tuple), and its null_value
member is accumulated over all elements of this tuple. So, as long as
the tuple contains a NULL element, the result of IN is set to UNKNOWN
by the first marked line, which is wrong.
Second fix: remove that line.

After that, another problem pops up (second marked line):
the "cmp" call is unaware of NULLs in the left argument (it has stored
only the "NULL-ignoring" value of this argument, for example the
return value of val_int (); leaving the handling of NULL to the first
marked line). So it is unable to calculate the proper value, because
it doesn't have the needed information.
This affects all descendent classes of cmp_item.
Third fix: first store NULLness information for _every_ element of the
left argument (each 'comparator[i]' of cmp_item_row), then make cmp()
use this information.

The result change in main.row is correct.