Bug#17720294 - TEST FROM INNODB.INNODB_MYSQL ASSERTS IN                PROTOCOL::END_STATEMENT
DESCRIPTION:

In the debug build the test case asserts for DELETE IGNORE
statement.

ANALYSIS:

The status of any statement execution is stored in Diagnostics
Area. Before ending an statement, the Diagnostics area is checked
for a valid status of the statement execution. After successful
execution of an statement, my_ok() function is called to set DA_OK
for the statement. Similarly, in case of a error, Diagnostics Area
is set with the code DA_ERROR.

When IGNORE keyword is used with the statement, the error is
suppressed. The statement should execute successfully. With multi
delete statement, error number is stored in an 'error' variable'.
This variable is used to check if my_ok() should be called.

With IGNORE keyword, multi delete statement suppress the error
reported by the storage engine during ha_delete_row() but error
code is retained in the 'error' variable. So, Diagnostics Area
is neither set to DA_OK or DA_ERROR which leads to an assert.

Fix:

Reset the 'error' variable in the multi delete if ignore is used.