Bug#17713388 RPL.RPL_SLAVE_LOAD_REMOVE_TMPFILE TEST IS FAILING ON WINDOWS ON PB2
Analysis: 
On a MTS enabled slave, in case a worker fails to
apply an event the slave applier (SQL thread +
workers) stops. An error that the worker faced is
printed in error log file. However the format of
the error in MTS enabled slave case is different
from a NON-MTS case( a prefix with worker id is
added)
Eg: Slave SQL: Worker 3 failed executing transaction 
The existing suppression in the mentioned test
script was not consider this prefix which gets added
in MTS enabled case

Fix: Changed the suppression to consider the prefix.