This is a patch for bug#17458917 - set_field_to_null, assertion failed: 0, file field_conv.cc, line 109
A crash was happened only against the server built in the debug mode.

The reason for the bug was that for the case when BLOB field is queried with
GROUP BY NULL and result is inserted into some table's column that declared
as NOT NULL then a Field object for such column doesn't have neither
is_tmp_nulable() on nor is_real_maybe_null() on. As a result when
set_field_to_null() is called the crash happened for the server built with
debug.

To fix the bug DBUG_ASSERT(0) was removed right before #ifdef DBUG_OFF 
and the #ifdef itselve was removed for the piece of code that catches
such a case. Actually, this patch reverts the changes introduced by wl#6030
for the function set_field_to_null(). Such changes were made since our mtr
test base doesn't cover the case when NULL BLOB column is fetched with
GROUP BY NULL and inserted into other table's column declared as NOT NULL.