Fix for bug#17738720:
  ARBITRATION TIMEOUT NOT INITIALIZED -> FALSE WARN 'COULD NOT FIND AN ARBITRATOR'
      
During arbitrator selection, QMGR runs through the states:
(NULL, INIT, FIND, PREP1, PREP2, START, ... and some more)
      
Currently the arbitration timeout timer is not set until we reach the
PREP-states. However, there is a check for timeout already in the
FIND-state which may then read uninitialized timestamp.
This may cause false warnings:
      
'Could not find an arbitrator, cluster is not partition-safe'
      
This fix sets the timeout timer in the 'INIT' state.