Bug#11762236 - optimizer use ref joins where it should use eq_ref
Problem: 
Optimize_table_order::find_best_ref() tries to pick the best index 
to access a table. If multiple indexes can be used, the choice of 
which index to pick is cost based. However, if a non-unique index 
has a rec_per_key estimate of one, the cost of using that index 
would be calculated to be the same as using EQ_REF access on a unique 
index even though EQ_REF guarantees "at most one match" semantics. 
This means that if the non-unique index is evaluated before the unique 
index, the unique index will not be chosen.

Note:
We should choose EQ_REF over REF, as we know that there will be 
no more than one matching row with EQ_REF while with REF the estimate 
may be completely wrong for this particular value.

Solution:
The best index is choosen using the following priority list
1) A clustered primary key is always chosen if all key parts have
   equality predicates.
2) A non nullable unique index with equality predicates on
   all keyparts is preferred over a non-unique index,
   nullable unique index or unique index where there are some
   keyparts without equality predicates.
3) Otherwise, the index with lowest cost estimate is chosen.