BUG 17447500 - FTS: RENAMING TO MISSING DATABASE:ASSERTION: TRX->STATE == TRX_STATE_ACTIVE
The missing database would result in the rename failure of aux tables. 
Then server will try to revert the rename. But the trx has already been
commited when server did the failed rename. So when the server calls
row_rename_table_for_mysql to revert the half done rename, the assertion,
which asserts the trx must be active, would be triggered.

If the rename fails, we have to allocate a new background trx to revert
the half done rename. As described in the comments, we don't have to deal
with the failure of the reverting.