Bug#17571371 : UNSTABLE THROUGHPUT FOR HUGE LOG_FILE_SIZE
New option innodb_log_write_ahead_size is added to avoid read-on-write for the transaction log file when not cached enough to the OS/FS cache.

innodb_log_write_ahead_size
  "Redo log write ahead unit size to avoid read-on-write,"
  " it should match the OS cache block IO size"

The write ahead should be cached to OS/FS and individual write log blocks in the cached area never cause read-on-write.

Its default value is 8K, because it is the block size at Solaris and larger than Linux default (4K). So enough for both cases.

Approved by Sunny in rb#3595