Bug #11758237 INSERT ON DUPLICATE KEY UPDATE SOMETIMES WRITES BINLOG POSITION INCORRECT
Problem:

When concurrent INSERT ... ON DUPLICATE UPDATE statement is executed on a
table, the final outcome is sometimes not serializable.

Solution:

When INSERT ... ON DUPLICATE UPDATE statement is execute, first the record
would be inserted into the clustered index followed by the secondary indexes.
If a duplicate error is encountered, we do not stop processing.  We continue to
process all the unique secondary indexes and place the necessary gap locks.
Once a duplicate error has been encountered, the records will not be actually
inserted but only gap locks will be placed.

rb#3196 approved by Marko