Bug#17746721 ASSERT IN THD::RESTORE_SUB_STATEMENT_STATE
The patch for bug#16041903 introduced a DBUG_ASSERT in
"THD::restore_sub_statement_state" just after
"in_sub_stmt" was reset to the value it had before
executing the sub statement.

The assert incorrectly assumed that
"is_fatal_sub_stmt_error" could be true only if we were
not in sub statement mode. However, since a sub statement
can call a new sub statement, there will be a sub statement
"execution stack", meaning that so the assert is wrong. 

The fix is to remove the incorrect assert. A new test is
added and put under the original bug number 16041903 since
it naturally belongs with the other test cases.