BUG 17607687 - FTS: DB STARTING WITH NUMBER: [FATAL] INNODB: PARSER: SYNTAX ERROR IN SQL STRING
If some special characters, such as numbers etc. are used as the
leading byte of a database name, the server will crash when we
access those fulltexts.

In fact this should be a duplicate bug with BUG#17607956 - FTS:
MORE SQL PARSER PROBLEMS, ASSERTION IN IN FILE PARS0LEX.L LINE 686

This patch is more on refactor than bug fix. It changes all those
\"%s\" to pars_info_bind_id, which is more safe and consistent
with the context.