Bug#14672988 : NEW BEHAVIOUR WHEN USING DISABLED ENGINE FAILS WARNINGS_ENGINE_DISABLED.TEST
Description:
-----------
The MySQL Server's behaviour how to handle usage of disabled engines has changed from MySQL-5.6 onwards. 
When a disabled engine is used to create a table, the server throws an error instead of a warning message.

Fix :
-----------
Added --error <error_message> in the test indicating that the error message is expected from the server.