BUG#17398792 - VALGRIND ERRORS IN THREAD_POOL.THREAD_POOL                CONNECTION_HANDLER PROBLEM AND FIX: The function net_write_raw_loop in the NET layer calls the thd_increment_bytes_sent. This function uses the current_thd macro which depends on pthread specific key THR_THD to get THD associated with the current connection. But before a connection context THD is initialized and hence THR_THD key is associated with a thread, errors like Out of memory can happen in which case we use send_error_and_close_channel of Channel_info object to propagate the error to the client.
  In the main process/thread that accepts the connection in
thread pool, thd is instantiated and initialized and the
connection-related contexts thd and thd->mem_root is
associated with the connection accept thread. Another
subsequent connection can fail before it's thd is
initialized in which case the accept thread may contain
stale thd of previous disconnected connection which
causes the valgrind failure.
  The fix is to diassociate the THD context when THD is
enqueued on to OS epoll wait queue to listen for query
command processing events.