Bug#17727506 REGRESSION: CRASH IN JOIN::RECOUNT_FIELD_TYPES
This is a regression introduced by the fix for bug #16967281.

Problem: Server may crash when executing an INSERT SELECT with UNION,
ROLLUP and ON DUPLICATE KEY UPDATE with a subquery.

When an INSERT SELECT statement has an ON DUPLICATE KEY UPDATE clause
that contains a subquery, the SELECT_LEX_UNIT for that subquery is
added as the slave of the last SELECT_LEX in the INSERT SELECT, even
though it is not part of the query itself. This subquery in the update
values list is not prepared before JOIN::optimize() on the INSERT
SELECT.

When JOIN::recount_field_types() tries to recursively call itself on
subqueries, it also traverses the SELECT_LEX for the update subquery,
which isn't prepared and doesn't have a JOIN object.

Fix: Don't try to recount field types in unprepared subqueries.