Bug #17233214 UNEXPECTED VALUE FOR SECONDS BEHIND MASTER
Problem:

Sporadically, SBM is 0 when it should be greater than 0.
Happens on 5.6+.

Analysis:

The issue is that exec_time on master server is being
replicated and used in the slave to calculate SBM.
Sometimes, that value is negative. After a couple of
iterations on analyzing this, what happened was that
"end_time" in log_event.cc used time() and thd->start_time
used my_micro_time() which have different resolutions and
API handling.

Fix:

The solution is to make end_time use my_micro_time(),
normalize it to seconds as it is been done in THD and then
calculate exec_time.