Fix for bug#17739131 JOB SCHEDULER IN SINGLE THREADED NDBD DOES NOT 'YIELD' CPU AS INTENDED
When the internal single threaded job scheduler is configured to use
'RealTimeScheduler=1', it is intended to temporary lower its scheduling
priority to 'normal' every 10ms. This will allow other, non realtime threads,
to get a chance to run, and thus avoid CPU starvation of these.
      
However, the arguments to the function which calculated elapsed time
singe last yield where swapped. Thus, '0' was always returned as elapsed
time from this function, and the CPU never yielded as intended.