MDEV-5182 - build of 10.0.4/r3863 fails @ 'cmake' with             -DINSTALL_SYSCONFDIR/-DDEFAULT_SYSCONFDIR specified
The reason for a bug is in support-files/CMakeLists.txt that tries to
install files in INSTALL_SYSCONF2DIR, without checking whether this
variable is actually defined. 

Normally INSTALL_SYSCONF2DIR is always defined if INSTALL_SYSCONFDIR is
defined, and support-files/CMakeLists.txt assumes that. But when one
specifies INSTALL_SYSCONFDIR manually on the command line, this
assumption might be broken.