Bug#17398972: SELECT statement causes 5.6.13 to crash
The problem occurs when Item_ident::fix_after_pullout()
encounters an Item_field object with context->select_lex equal to
depended_from during semi-join transformation.

This object was created when resolving a column from the HAVING clause,
which was found in a table from the middle query block (it is found
in the inner-most query block too, but since it is inside HAVING and
not in the SELECT list nor in GROUP BY, it is resolved one step out).

Inside Item_ref::fix_fields(), a new Item_field object is created.
But unlike almost every other Item_field/Item_ref object, it is
created with context pointing to the query block where the column is
resolved from, and not the query block in which the column reference
is placed.

The fix is to change this context to the query block where the column
reference is placed.