Bug#17813333 PSI KEYS WERE INCORRECTLY PASSED TO FUNCTION SET_PSI_KEYS
Before this fix, the mutex 
  'wait/synch/mutex/sql/MYSQL_RELAY_LOG::LOCK_sync'
was not properly instrumented for the performance schema.

In particular, no wait time was recorded for this mutex,
and no mutex instance was instrumented,
because the incorrect key was used when instrumenting the code.

The root cause is a simple mismatch in the parameters order,
in the call to relay_log.set_psi_keys().

This fix uses the proper keys when calling set_psi_keys()
for the relay log.