Bug#70484/17547804: Creating view will crash server without error message
This is a regression from the refactoring part of bug#16436383.

The problem is that when resolving aggregation expressions,
we follow the chain of outer query blocks to locate the appropriate
query block for aggregation of the expression. This works fine in
the general case, but when doing this for a view that has been
expanded into an outer query, we may actually follow the chain
of outer query blocks into the outer query. The reason for this is
that the outer-most query block of the view has nest level 0, and
we are not prevented from exiting the loop in this situation.

The solution here is to explicitly exit the loop when nest level is 0.

Note that this is a problem only in 5.6. In 5.7, WL#4772 changed
numbering of nesting levels when attaching view queries to the
outer query so that the outer-most query block of the view got
nest level 1 (instead of 0) and the inner query block got nest level
2 (instead of 1). In this case, the algorithm from bug#16436383
works properly, so this fix is not needed there.

Note that the max_arg_level member follows the nest levels, pointing
to the first available nest level bit in allow_sum_func, ie 0 in
5.6 and 1 in 5.7.