Part2 of fix (the real fix) for bug#17647637:
'NDBTICK_CURRENTMICROSECOND() MUST BE MONOTONIC'
  
Use 'QueryPerformanceCounters' on Windows instead of GetSystemTimeAsFileTime()
to implement the NdbTick timers. This will ensure that the Windows timers are 
monotonic.
   
Also removes the 'need_monotonic' argument for NdbTick_Init() and
NdbCondition_initialize(). Now  we always tries to use a monotonic
timer, and print a warning if this was not possible.

Furthermore, CLOCK_HIGHRES is introduced as an alternative for 
CLOCK_MONOTONIC if that is not available on the platform (Solaris 9).

NdbTick.cpp was refactored such that there are not two platform dependent
versions of NdbTick_Init() and NdbTick_getCurrentTicks() any more.