bug#17703157  HAVE_BINLOG_ROWS_QUERY.INC : IMPLEMENTATION LOGIC DOES NOT MATCH PURPOSE
The 'have_binlog_rows_query.inc' and 'have_binlog_rows_query.require'
are introduced to ensure that the server is running with
binlog_rows_query_log_events on for rpl_row_ignorable_event.test.
However, the current implementation is as following: 
  1) Fetch variable binlog_rows_query_log_events
  2) Replace ON by OFF in the result
  3) Check if result value is OFF
The purpose and implementation are not aligned.
  
Actually the 'have_binlog_rows_query.inc' and 'have_binlog_rows_query.require'
are not necessary as we already enabled binlog-rows-query-log-events option
in rpl_row_ignorable_event-master.opt. So we remove the redundent two files.