Removing the usuage of global variable server_version from libbinlogapi
Global variable server_version was declared in binlog_event.h and defined in 
the server, this introduces a dependency on the server for libbinlogapi.
Now we are passing server_version to the ctor of Format_description_event.
The ctor of Format_description_event is changed a little, now there is no default
value for server_version, which will make sure that
we dont miss passing server_version.

And we have added one more parameter to the ctor of 
Binary_log_event(const char *buf, uint16_t binlog_version, const char *server_version)