Bug#17543588 PERFORMANCE REGRESSION (9%) FOR DBT-3 QUERY 21
This is an unexpected regression from an earlier fix:

Bug#16852278 SIMPLIFY RECORD COMPARISONS

The reason of this regression is that memcmp() can be slower than a for loop.
If memcmp() is a library call, there is an overhead involved,
which we can avoid by comparing the first few bytes with a loop.
The built-in memcmp() of GCC on x86 and AMD64 (repz cmpsb) is known to
perform worse than the GNU libc library function:

http://gcc.gnu.org/bugzilla/show_bug.cgi?id=43052

cmp_data(): Use a for loop to compare some first bytes.
Then invoke memcmp() for any remaining common bytes if needed.

innodb.cmake: Pass -fno-builtin-memcmp when compiling rem0cmp.cc
with GCC on x86 or AMD64.

rb#3931 approved by Jimmy Yang