Bug#17774971 MOVE "THD_MARK_TRANSACTION_FOR_ROLLBACK" DEFINITION FROM HA_INNODB.H TO PLUGIN.H
A post push fix for bug#16041903 modified ha_ndbcluster.cc to
call THD::mark_transaction_to_rollback instead. The Cluster
team requested that the function should be defined as part of
the API in include/mysql/plugin.h. 

There is a corresponding function in sql/sql_class.cc. The
function is used by InnoDB, and is defined in ha_innodb.h.

Thus, the suggested refactoring is to:

1. Move the definition of thd_mark_transaction_to_rollback
   from ha_innodb.h to plugin.h
2. Change the type of the paramter "all" from bool to int,
   because it is now defined in plugin.h which is compiled
   as both C and C++, and the type "my_bool" is not available
   in plugin.h
3. Change ha_ndbcluster.cc to call the above function.
4. Change ha_innodb.cc to submit int rather than ibool for
   the "all" parameter
5. Change the implemntation in sql_class.cc to test for
   (all != 0) explicitly when calling the THD member
   function