Bug#17309657 UPDATE ON P_S SETUP_CONSUMERS THROUGH JOIN ONLY UPDATES FIRST ROW
Before this fix, complex updates to performance schema tables,
when involving joins / sub queries, failed to update every row as expected.

The root cause is that a storage engine implementation is supposed to
initialize the table->status field every time a row is scanned,
but the performance_schema failed to do so.

table->status is used in particular in multi_update::send_data()
to keep track of which rows have been updated already.
The lack of initialization of table->status for each new row confused this
code, causing the bug found.

The fix is to initialize table->status in ha_perfschema::rnd_next() and
ha_perfschema::rnd_pos(), in a manner consistent with exiting storage engines
such as innodb or myisam.